Now supports closed-source copyrighted Google Play Location Services (aka GMS) if available, else falls back to the old-style Android Open Source Project way of doing things.
git-svn-id: http://little-fluffy-location-library.googlecode.com/svn@16 88fb8fb9-40e5-4f22-20ed-f1c1febcbddd

+++ trunk/ little-fluffy-location-library/LittleFluffyLocationLibrary/src/com/littlefluffytoys/littlefluffylocationlibrary/LocationBroadcastService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012 Little Fluffy Toys Ltd
+ * Copyright 2014 Little Fluffy Toys Ltd
  * Adapted from work by Reto Meier, Copyright 2011 Google Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -20,6 +20,14 @@ package com.littlefluffytoys.littlefluffylocationlibrary;
 import java.util.Timer;
 import java.util.TimerTask;
 
+import com.google.android.gms.common.ConnectionResult;
+import com.google.android.gms.common.GooglePlayServicesUtil;
+import com.google.android.gms.common.GooglePlayServicesClient.ConnectionCallbacks;
+import com.google.android.gms.common.GooglePlayServicesClient.OnConnectionFailedListener;
+import com.google.android.gms.location.LocationClient;
+import com.google.android.gms.location.LocationListener;
+import com.google.android.gms.location.LocationRequest;
+
 import android.annotation.TargetApi;
 import android.app.AlarmManager;
 import android.app.PendingIntent;
@@ -30,9 +38,9 @@ import android.content.SharedPreferences;
 import android.content.SharedPreferences.Editor;
 import android.location.Criteria;
 import android.location.Location;
-import android.location.LocationListener;
 import android.location.LocationManager;
 import android.os.Binder;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.IBinder;
 import android.os.Parcel;
@@ -48,9 +56,12 @@ import android.util.Log;
  * @see AlarmService
  * @see AlarmService_Alarm
  */
-public class LocationBroadcastService extends Service {
+public class LocationBroadcastService extends Service implements ConnectionCallbacks, OnConnectionFailedListener, LocationListener {
     
     private static final String TAG = "LocationBroadcastService"; 
+    
+	private LocationClient mLocationClient;
+	private LocationRequest mLocationRequest;
 
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
@@ -74,7 +85,8 @@ public class LocationBroadcastService extends Service {
      * The function that runs in our worker thread
      */
     Runnable mTask = new Runnable() {
-        public void run() {
+        @TargetApi(Build.VERSION_CODES.GINGERBREAD)
+		public void run() {
             boolean stopServiceOnCompletion = true;
 
             final SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
@@ -88,7 +100,14 @@ public class LocationBroadcastService extends Service {
 
                 if (forceLocationToUpdate || System.currentTimeMillis() - lastLocationUpdateTimestamp > LocationLibrary.getLocationMaximumAge()) {
                     if (forceLocationToUpdate) {
-                        prefs.edit().putBoolean(LocationLibraryConstants.SP_KEY_FORCE_LOCATION_UPDATE, false).commit();
+                        final Editor prefsEditor = prefs.edit();
+                        prefsEditor.putBoolean(LocationLibraryConstants.SP_KEY_FORCE_LOCATION_UPDATE, false);
+                        if (LocationLibraryConstants.SUPPORTS_GINGERBREAD) {
+                        	prefsEditor.apply();
+                        }
+                        else {
+                        	prefsEditor.commit();
+                        }
                     }
                     // Current location is out of date. Force an update, and stop service if required.
                     stopServiceOnCompletion = !forceLocationUpdate();
@@ -101,7 +120,12 @@ public class LocationBroadcastService extends Service {
                 if (forceLocationToUpdate) {
                     prefsEditor.putBoolean(LocationLibraryConstants.SP_KEY_FORCE_LOCATION_UPDATE, false);
                 }
-                prefsEditor.commit();
+                if (LocationLibraryConstants.SUPPORTS_GINGERBREAD) {
+                	prefsEditor.apply();
+                }
+                else {
+                	prefsEditor.commit();
+                }
                 sendBroadcast(getBaseContext(), true);
             }
 
@@ -135,31 +159,39 @@ public class LocationBroadcastService extends Service {
         final Criteria criteria = new Criteria();
         criteria.setAccuracy(!LocationLibrary.useFineAccuracyForRequests && locationManager.isProviderEnabled(LocationManager.NETWORK_PROVIDER) ? Criteria.ACCURACY_COARSE : Criteria.ACCURACY_FINE);
 
-        if (LocationLibraryConstants.SUPPORTS_GINGERBREAD) {
-            if (LocationLibrary.showDebugOutput) Log.d(LocationLibraryConstants.TAG, TAG + ": Force a single location update, as current location is beyond the oldest location permitted");
-            // just request a single update. The passive provider will pick it up.
-            final Intent receiver = new Intent(getApplicationContext(), PassiveLocationChangedReceiver.class).addCategory(LocationLibraryConstants.INTENT_CATEGORY_ONE_SHOT_UPDATE);
-            final PendingIntent oneshotReceiver = PendingIntent.getBroadcast(getApplicationContext(), 0, receiver, PendingIntent.FLAG_UPDATE_CURRENT);
-            try {
-                locationManager.requestSingleUpdate(criteria, oneshotReceiver);
-                if (LocationLibrary.showDebugOutput) Log.d(LocationLibraryConstants.TAG, TAG + ": schedule timer to kill locationlistener in 30 seconds");
-                new Timer().schedule(new TimerTask(){
-                    public void run(){
-                        try {
-                            if (LocationLibrary.showDebugOutput) Log.d(LocationLibraryConstants.TAG, TAG + ": remove updates after 30 seconds");
-                            locationManager.removeUpdates(oneshotReceiver);
-                        }
-                        catch (Exception e) {
-                            e.printStackTrace();
-                        }
-                        stopSelf();
-                    }}, 30000);
-                return true; // don't stop the service, allow the timer to do that
-            }
-            catch (IllegalArgumentException ex) {
-                // thrown if there are no providers, e.g. GPS is off
-                if (LocationLibrary.showDebugOutput) Log.w(LocationLibraryConstants.TAG, TAG + ": IllegalArgumentException during call to locationManager.requestSingleUpdate - probable cause is that all location providers are off. Details: " + ex.getMessage());
-            }
+	    if (LocationLibraryConstants.SUPPORTS_GINGERBREAD) {
+	    	if (GooglePlayServicesUtil.isGooglePlayServicesAvailable(getApplicationContext()) == ConnectionResult.SUCCESS) {
+	            if (LocationLibrary.showDebugOutput) Log.d(LocationLibraryConstants.TAG, TAG + ": Force a single location update using Google GMS Location, as current location is beyond the oldest location permitted");
+		    	mLocationClient = new LocationClient(getApplicationContext(), this, this);
+	    		mLocationClient.connect(); // this will cause an onConnected() or onConnectionFailed() callback 
+	    		return true;
+		    }
+		    else {
+	            if (LocationLibrary.showDebugOutput) Log.d(LocationLibraryConstants.TAG, TAG + ": Force a single location update using Android AOSP location, as current location is beyond the oldest location permitted");
+	            // just request a single update. The passive provider will pick it up.
+	            final Intent receiver = new Intent(getApplicationContext(), PassiveLocationChangedReceiver.class).addCategory(LocationLibraryConstants.INTENT_CATEGORY_ONE_SHOT_UPDATE);
+	            final PendingIntent oneshotReceiver = PendingIntent.getBroadcast(getApplicationContext(), 0, receiver, PendingIntent.FLAG_UPDATE_CURRENT);
+	            try {
+	                locationManager.requestSingleUpdate(criteria, oneshotReceiver);
+	                if (LocationLibrary.showDebugOutput) Log.d(LocationLibraryConstants.TAG, TAG + ": schedule timer to kill locationlistener in 30 seconds");
+	                new Timer().schedule(new TimerTask(){
+	                    public void run(){
+	                        try {
+	                            if (LocationLibrary.showDebugOutput) Log.d(LocationLibraryConstants.TAG, TAG + ": remove updates after 30 seconds");
+	                            locationManager.removeUpdates(oneshotReceiver);
+	                        }
+	                        catch (Exception e) {
+	                            e.printStackTrace();
+	                        }
+	                        stopSelf();
+	                    }}, 30000);
+	                return true; // don't stop the service, allow the timer to do that
+	            }
+	            catch (IllegalArgumentException ex) {
+	                // thrown if there are no providers, e.g. GPS is off
+	                if (LocationLibrary.showDebugOutput) Log.w(LocationLibraryConstants.TAG, TAG + ": IllegalArgumentException during call to locationManager.requestSingleUpdate - probable cause is that all location providers are off. Details: " + ex.getMessage());
+	            }
+		    }
         }
         else { // pre-Gingerbread
             if (LocationLibrary.showDebugOutput) Log.d(LocationLibraryConstants.TAG, TAG + ": Force location updates (pre-Gingerbread), as current location is beyond the oldest location permitted");
@@ -185,7 +217,7 @@ public class LocationBroadcastService extends Service {
         alarm.set(AlarmManager.RTC_WAKEUP, System.currentTimeMillis() + (delayInSeconds * 1000), pIntent);
     }
 
-    final LocationListener preGingerbreadUpdatesListener = new LocationListener() {
+    final android.location.LocationListener preGingerbreadUpdatesListener = new android.location.LocationListener() {
         public void onLocationChanged(Location location) {
             if (LocationLibrary.showDebugOutput) Log.d(LocationLibraryConstants.TAG, TAG + ": Single Location Update Received: " + location.getLatitude() + "," + location.getLongitude());
             ((LocationManager) getApplicationContext().getSystemService(Context.LOCATION_SERVICE)).removeUpdates(preGingerbreadUpdatesListener);
@@ -216,4 +248,66 @@ public class LocationBroadcastService extends Service {
             return super.onTransact(code, data, reply, flags);
         }
     };
+    
+	/**
+	 * Google GMS Location
+	 */
+	public void onConnectionFailed(ConnectionResult arg0)
+	{
+        if (LocationLibrary.showDebugOutput) Log.w(LocationLibraryConstants.TAG, TAG + ": onConnectionFailed (Google GMS Location)");
+        stopSelf();
+	}
+
+	/**
+	 * Google GMS Location
+	 */
+	public void onConnected(Bundle arg0)
+	{
+        if (LocationLibrary.showDebugOutput) Log.d(LocationLibraryConstants.TAG, TAG + ": onConnected (Google GMS Location)");
+        mLocationRequest = LocationRequest.create();
+        mLocationRequest.setPriority(LocationLibrary.useFineAccuracyForRequests ? LocationRequest.PRIORITY_HIGH_ACCURACY : LocationRequest.PRIORITY_BALANCED_POWER_ACCURACY);
+        mLocationRequest.setInterval(4000);
+        mLocationRequest.setFastestInterval(1000);
+        mLocationClient.requestLocationUpdates(mLocationRequest, LocationBroadcastService.this);
+	}
+
+	/**
+	 * Google GMS Location
+	 */
+	public void onDisconnected()
+	{
+        if (LocationLibrary.showDebugOutput) Log.d(LocationLibraryConstants.TAG, TAG + ": onDisconnected (Google GMS Location)");
+	}
+
+	/**
+	 * Google GMS Location
+	 */
+	public void onLocationChanged(Location location)
+	{
+        if (LocationLibrary.showDebugOutput) Log.d(LocationLibraryConstants.TAG, TAG + ": onLocationChanged (Google GMS Location)");
+		mLocationClient.removeLocationUpdates(this);
+		mLocationClient.disconnect();
+		PassiveLocationChangedReceiver.processLocation(LocationBroadcastService.this, location, false, true);
+	}
+
+	/**
+	 * Google GMS Location
+	 */
+	public void onProviderDisabled(String provider)
+	{
+	}
+
+	/**
+	 * Google GMS Location
+	 */
+	public void onProviderEnabled(String provider)
+	{
+	}
+
+	/**
+	 * Google GMS Location
+	 */
+	public void onStatusChanged(String provider, int status, Bundle extras)
+	{
+	}
 }
